- Ajout des livreurs d'eau et des gardes de sécurité (IA simple) - Ajout de la gestion de destruction des batiments - Ajout de la route bloquante

+++ src/com/exiro/render/MouseManager.java
@@ -46,7 +46,8 @@ public class MouseManager implements MouseListener {
         }
         if (e.getButton() == MouseEvent.BUTTON3) {
             if (build) {
-                win.p.getPlayerCities().get(0).getMap().getCase(c.getxPos(), c.getyPos()).getObject().delete();
+                if (win.p.getPlayerCities().get(0).getMap().getCase(c.getxPos(), c.getyPos()).getObject() != null)
+                    win.p.getPlayerCities().get(0).getMap().getCase(c.getxPos(), c.getyPos()).getObject().delete();
             }
         }
